More unsigned 32 bit integer fixes.
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@1102 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ org/xbill/DNS/Tokenizer.java
@@ -460,7 +460,7 @@ getDouble() throws IOException {
  * @throws IOException An I/O error occurred.
  * @see TTL
  */
-public int
+public long
 getTTL() throws IOException {
 	String next = getIdentifier();
 	try {
